+++ examples/http_server_usage/datetime_example.py
@@ -16,7 +16,7 @@ import pyjsonrpc.rpcjson
 
 
 # Activate encoder and decoder for DateTime-ISO strings and NDB keys
-pyjsonrpc.rpcjson.activate_date_time_and_ndb_conversion()
+pyjsonrpc.rpcjson.activate_iso_date_and_ndb_conversion()
 
 
 class RequestHandler(pyjsonrpc.HttpRequestHandler):